plugin-manager/requirements.txt
coverage==7.3.0
cryptography>=41.0.4
docker==7.1.0
fastapi[all]~=0.115.2
GitPython==3.1.41
httpx>=0.25.0
minio==7.2.4
numpy==1.26.4
paho-mqtt==1.6.1
pandas==2.2.2
prometheus-fastapi-instrumentator==7.0.0
pycryptodomex>=3.19.0
pydantic==2.7.3
PyJWT==2.8.0
pymongo==4.6.3
pytest==7.4.2
pytest-asyncio==0.23.8
pytest-dotenv==0.5.2
pytest-mock==3.12.0
python-dotenv~=1.0.0
python-multipart==0.0.18
redis==4.6.0
shortuuid~=1.0.11
ujson==5.8.0
urllib3~=2.4.0
ut-notifications-util==0.0.7.04
ut-security-util==1.2.2
uvicorn[standard]>=0.23.2
